Add files via upload
Added a link to the GitHub page in the navigation bar, changed the layout of the file explorer, and fixed a very serious bug that would delete the user folder if no files were selected.

+++ delete.php
@@ -4,37 +4,40 @@
 	echo $locationfile;
 	echo $d;
 */
+	if (!empty($_POST["file"])) {
+		session_start();
 
-	session_start();
+		var_dump($_POST["file"]);
 
-	var_dump($_POST["file"]);
+		// récupére le chemin du dossier actuel (celui de ce script)
+		$path = dirname(__FILE__).'/';
 
-	// récupére le chemin du dossier actuel (celui de ce script)
-	$path = dirname(__FILE__).'/';
+		// affiche le contenu de la session
+		var_dump($_SESSION);
 
-	// affiche le contenu de la session
-	var_dump($_SESSION);
+		$locationfile = './'.$_SESSION["user"].'/'.$_POST["file"];
 
-	$locationfile = './'.$_SESSION["user"].'/'.$_POST["file"];
+		var_dump($locationfile);
 
-	var_dump($locationfile);
+		// test si le dossier existe
+		if (!file_exists($locationfile)) {
+			var_dump('Introuvable');
+			// tue le script
+			die();
+		}
 
-	// test si le dossier existe
-	if (!file_exists($locationfile)) {
-		var_dump('Introuvable');
-		// tue le script
-		die();
+	
+		if (is_dir($_SESSION["user"]."/".$_POST["file"])) {
+			rmdir ($_SESSION["user"]."/".$_POST["file"]);
+		}
+		else {
+			unlink($_SESSION["user"]."/".$_POST["file"]);
+		}
+
+		
 	}
-
-	if (is_dir($_SESSION["user"]."/".$_POST["file"])) {
-		rmdir ($_SESSION["user"]."/".$_POST["file"]);
-	}
-	else {
-		unlink($_SESSION["user"]."/".$_POST["file"]);
-	}
-
 	// header() renvoit sur une autre page
-    header('Location: singed.php');
+	header('Location: singed.php');
 ?>
 
 <!--
